pyproject.toml
[project]
name = "crm-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=37.3.0",
    "langchain-core>=0.3.62",
    "langchain-mcp-adapters>=0.1.1",
    "langchain-openai>=0.3.18",
    "langgraph>=0.4.7",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.0.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pandas>=2.2.3",
]
